Add RC augmentation
