+++ experiments/plantcad/exp2101_plantcad_isoflop_analysis.py
@@ -14,6 +14,7 @@
 
 import argparse
 import json
+import logging
 from pathlib import Path
 
 import matplotlib.pyplot as plt
@@ -24,9 +25,50 @@ from rich.table import Table
 from scipy.interpolate import griddata
 import wandb
 
-RUN_PREFIX = "plantcad_isoflop_v1.0"
-RESULT_PATH = "experiments/plantcad/results/v3"
+RUN_PREFIX = "plantcad_isoflop_v1.1"
+RESULT_PATH = "experiments/plantcad/results/v4"
 EXPORT_DPI = 300
+DEFAULT_ARCH = "qwen"
+
+console = Console(record=True)
+logger = logging.getLogger(__name__)
+
+
+def setup_logging(log_path: Path) -> None:
+    """Configure logging to both console and file."""
+    log_path.parent.mkdir(parents=True, exist_ok=True)
+
+    # Clear any existing handlers
+    logger.handlers.clear()
+    logger.setLevel(logging.INFO)
+
+    # Console handler
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(logging.INFO)
+    console_handler.setFormatter(logging.Formatter("%(message)s"))
+
+    # File handler
+    file_handler = logging.FileHandler(log_path, mode="w")
+    file_handler.setLevel(logging.INFO)
+    file_handler.setFormatter(logging.Formatter("%(asctime)s - %(message)s", datefmt="%Y-%m-%d %H:%M:%S"))
+
+    logger.addHandler(console_handler)
+    logger.addHandler(file_handler)
+
+
+def filter_to_finished_runs(df: pd.DataFrame, allow_crashed: bool = False) -> pd.DataFrame:
+    """Filter dataframe to include finished runs and nearly-complete crashed runs.
+
+    Includes runs where:
+    - state == "finished", OR
+    - state == "crashed" AND run_progress > 0.999
+    """
+    is_finished = df["state"] == "finished"
+    if allow_crashed:
+        is_nearly_complete_crash = (df["state"] == "crashed") & (df["run_progress"] > 0.999)
+        return df[is_finished | is_nearly_complete_crash]
+    else:
+        return df[is_finished]
 
 
 def save_figure(fig, output_path: str) -> None:
@@ -36,19 +78,19 @@ def save_figure(fig, output_path: str) -> None:
 
     # Save PNG
     fig.savefig(output_path, dpi=EXPORT_DPI, bbox_inches="tight")
-    print(f"Saved plot to {output_path}")
+    logger.info(f"Saved plot to {output_path}")
 
     # Save PDF
     pdf_path = output_path.with_suffix(".pdf")
     fig.savefig(pdf_path, dpi=EXPORT_DPI, bbox_inches="tight")
-    print(f"Saved plot to {pdf_path}")
+    logger.info(f"Saved plot to {pdf_path}")
 
 
 def log_run_object(run, run_idx):
     """Log a run object as JSON to show available data."""
-    print(f"\n{'=' * 80}")
-    print(f"RUN {run_idx + 1}: {run.name}")
-    print(f"{'=' * 80}")
+    logger.info(f"\n{'=' * 80}")
+    logger.info(f"RUN {run_idx + 1}: {run.name}")
+    logger.info(f"{'=' * 80}")
     run_dict = {
         "id": run.id,
         "name": run.name,
@@ -58,8 +100,8 @@ def log_run_object(run, run_idx):
         "config": dict(run.config),
         "summary": dict(run.summary),
     }
-    print(json.dumps(run_dict, indent=2, default=str))
-    print(f"{'=' * 80}\n")
+    logger.info(json.dumps(run_dict, indent=2, default=str))
+    logger.info(f"{'=' * 80}\n")
 
 
 def fetch_plantcad_runs(show_wandb_runs: bool = False):
@@ -140,7 +182,7 @@ def save_runs(df, output_path=f"{RESULT_PATH}/plantcad_isoflops.csv"):
     output_path = Path(output_path)
     output_path.parent.mkdir(parents=True, exist_ok=True)
     df.to_csv(output_path, index=False)
-    print(f"Saved {len(df)} runs to {output_path}")
+    logger.info(f"Saved {len(df)} runs to {output_path}")
 
 
 def validate_runs(df):
@@ -148,15 +190,14 @@ def validate_runs(df):
     key_cols = ["architecture", "flops_budget", "tokens", "params", "epochs"]
     duplicates = df[df.duplicated(subset=key_cols, keep=False)]
     if not duplicates.empty:
-        print(f"WARNING: Found {len(duplicates)} duplicate rows by {key_cols}:")
-        print(duplicates[["run_name", *key_cols]].to_string())
+        logger.warning(f"Found {len(duplicates)} duplicate rows by {key_cols}:")
+        logger.warning(duplicates[["run_name", *key_cols]].to_string())
     else:
-        print(f"Validation passed: rows are unique by {key_cols}")
+        logger.info(f"Validation passed: rows are unique by {key_cols}")
 
 
 def summarize_runs(df):
     """Print formatted summary tables using rich."""
-    console = Console()
     gflops_to_flops = 1e9
 
     # Simplified summary table
@@ -195,10 +236,10 @@ def summarize_runs(df):
 def visualize_loss_by_token_count(df, metric="eval_loss", output_path=f"{RESULT_PATH}/plantcad_loss_by_tokens.png"):
     """Plot loss vs tokens, colored by budget, faceted by architecture (cols) and epochs (rows)."""
     required_cols = [metric, "tokens", "architecture", "flops_budget", "epochs"]
-    df_clean = df[df["state"] == "finished"].dropna(subset=required_cols)
+    df_clean = filter_to_finished_runs(df).dropna(subset=required_cols)
 
     if df_clean.empty:
-        print(f"Warning: No finished runs with required columns {required_cols}. Skipping visualization.")
+        logger.warning(f"No finished runs with required columns {required_cols}. Skipping visualization.")
         return
 
     architectures = sorted(df_clean["architecture"].unique())
@@ -258,10 +299,10 @@ def visualize_loss_by_token_count(df, metric="eval_loss", output_path=f"{RESULT_
 def visualize_loss_by_param_count(df, metric="eval_loss", output_path=f"{RESULT_PATH}/plantcad_loss_by_params.png"):
     """Plot loss vs params, colored by budget, faceted by architecture (cols) and epochs (rows)."""
     required_cols = [metric, "params", "architecture", "flops_budget", "epochs"]
-    df_clean = df[df["state"] == "finished"].dropna(subset=required_cols)
+    df_clean = filter_to_finished_runs(df).dropna(subset=required_cols)
 
     if df_clean.empty:
-        print(f"Warning: No finished runs with required columns {required_cols}. Skipping visualization.")
+        logger.warning(f"No finished runs with required columns {required_cols}. Skipping visualization.")
         return
 
     architectures = sorted(df_clean["architecture"].unique())
@@ -318,135 +359,6 @@ def visualize_loss_by_param_count(df, metric="eval_loss", output_path=f"{RESULT_
     save_figure(fig, output_path)
 
 
-def format_large_num(n):
-    """Simple formatter for large numbers (e.g. 1.5B, 100M)."""
-    if n is None:
-        return "N/A"
-    try:
-        n = float(n)
-    except (ValueError, TypeError):
-        return str(n)
-    if n >= 1e9:
-        return f"{n / 1e9:.1f}B"
-    if n >= 1e6:
-        return f"{n / 1e6:.1f}M"
-    if n >= 1e3:
-        return f"{n / 1e3:.1f}K"
-    return f"{n:.0f}"
-
-
-HTML_BASE_CSS = """
-<style type="text/css">
-table.dataframe {
-  border-collapse: collapse;
-  border: 1px solid #aaa;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-  font-size: 11px;
-}
-table.dataframe th, table.dataframe td {
-  border: 1px solid #aaa;
-  padding: 4px 6px;
-}
-table.dataframe thead tr {
-  background-color: #f0f0f0;
-}
-table.dataframe th.row_heading, table.dataframe th.index_name {
-  text-align: left;
-  white-space: nowrap;
-}
-table.dataframe td {
-  text-align: right;
-}
-</style>
-"""
-
-
-def visualize_loss_by_epochs(
-    df,
-    metric: str = "eval_loss",
-    output_path: str = f"{RESULT_PATH}/plantcad_loss_by_epochs.html",
-) -> None:
-    """Export loss vs. epochs as an HTML table with row-wise color gradients."""
-    required_cols = [metric, "tokens", "params", "architecture", "flops_budget", "epochs"]
-    df_clean = df[df["state"] == "finished"].dropna(subset=required_cols).copy()
-
-    architectures = sorted(df_clean["architecture"].unique())
-    if not architectures:
-        print("No data to visualize.")
-        return
-
-    # If the user passed a PNG path, switch to HTML next to it.
-    out_path = Path(output_path)
-    if out_path.suffix.lower() != ".html":
-        out_path = out_path.with_suffix(".html")
-
-    html_sections: list[str] = []
-
-    for arch in architectures:
-        df_arch = df_clean[df_clean["architecture"] == arch].copy()
-        if df_arch.empty:
-            continue
-
-        # Format columns for display in the index
-        df_arch["budget_str"] = df_arch["flops_budget"].apply(lambda x: f"{x:.1e}")
-        df_arch["params_str"] = df_arch["params"].apply(format_large_num)
-        df_arch["tokens_str"] = df_arch["tokens"].apply(format_large_num)
-
-        # Pivot: Index=(Budget, Params, Tokens), Columns=Epochs, Values=Loss
-        pivot = df_arch.pivot_table(
-            index=["flops_budget", "budget_str", "params", "params_str", "tokens", "tokens_str"],
-            columns="epochs",
-            values=metric,
-        )
-        if pivot.empty:
-            continue
-        # Sort rows: high compute first, then params, then tokens
-        pivot = pivot.sort_index(level=["flops_budget", "params", "tokens"], ascending=[False, True, True])
-
-        # Replace MultiIndex with a single string index using C, N, D labels
-        def _format_index(idx, pivot=pivot) -> str:
-            names = pivot.index.names
-            c = idx[names.index("budget_str")]
-            n = idx[names.index("params_str")]
-            d = idx[names.index("tokens_str")]
-            return f"C: {c} | N: {n} | D: {d}"
-
-        pivot.index = pd.Index([_format_index(idx) for idx in pivot.index], name=None)
-
-        # Make column labels nicer (e.g., "1 Ep") and ensure no extra header rows for column names
-        pivot.columns = [f"{int(col)} Ep" if float(col).is_integer() else str(col) for col in pivot.columns]
-        pivot.columns.name = None
-
-        # Style similar to pandas Styler in a Jupyter notebook:
-        #  - row-wise background gradient so each configuration uses full cmap span
-        #  - minimal extra styling, just number formatting and a caption.
-        styled = (
-            pivot.style.format("{:.4f}")
-            .background_gradient(axis=1, cmap="RdYlGn_r")
-            .set_table_attributes('class="dataframe"')
-            .set_caption(
-                f"Architecture: {arch} "
-                "(C = compute budget, N = params, D = tokens; colors are row-wise normalized loss)"
-            )
-        )
-
-        html_sections.append(styled.to_html())
-
-    if not html_sections:
-        print("No tables generated for any architecture.")
-        return
-
-    full_html = (
-        "<html><head><meta charset='utf-8'>"
-        "<title>PlantCAD Loss by Epochs</title>"
-        f"{HTML_BASE_CSS}</head><body>\n" + "<br><br>\n".join(html_sections) + "\n</body></html>"
-    )
-
-    out_path.parent.mkdir(parents=True, exist_ok=True)
-    out_path.write_text(full_html, encoding="utf-8")
-    print(f"Saved HTML table visualization to {out_path}")
-
-
 def get_size_label(rank: int) -> str:
     """Map model size rank to human-readable label."""
     labels = ["XXS", "XS", "S", "M", "L", "XL", "XXL", "XXXL"]
@@ -455,19 +367,19 @@ def get_size_label(rank: int) -> str:
     return f"Size-{rank}"
 
 
-def visualize_loss_by_epochs_matplotlib(
+def visualize_loss_by_epochs(
     df,
     metric: str = "eval_loss",
     output_path: str = f"{RESULT_PATH}/plantcad_loss_by_epochs.png",
 ) -> None:
     """Plot normalized loss vs epochs, faceted by architecture and budget."""
     required_cols = [metric, "tokens", "params", "architecture", "flops_budget", "epochs"]
-    df_clean = df[df["state"] == "finished"].dropna(subset=required_cols).copy()
+    df_clean = filter_to_finished_runs(df).dropna(subset=required_cols).copy()
 
     architectures = sorted(df_clean["architecture"].unique())
     budgets = sorted(df_clean["flops_budget"].unique(), reverse=True)
     if not architectures or not budgets:
-        print("No data to visualize.")
+        logger.warning("No data to visualize.")
         return
 
     # Normalize loss to 0-1 per (arch, budget, tokens, params) group
@@ -533,17 +445,17 @@ def visualize_loss_by_epochs_matplotlib(
 
 def visualize_loss_by_param_and_epoch_count(
     df,
-    architecture: str = "qwen",
+    architecture: str = DEFAULT_ARCH,
     metric: str = "eval_loss",
     clip_percentile: float = 80.0,
     output_path: str = f"{RESULT_PATH}/plantcad_loss_contour.png",
 ) -> None:
     """2D contour plot of loss vs params (y) and epochs (x), faceted by flops budget."""
     required_cols = [metric, "params", "epochs", "architecture", "flops_budget"]
-    df_clean = df[(df["state"] == "finished") & (df["architecture"] == architecture)].dropna(subset=required_cols).copy()
+    df_clean = filter_to_finished_runs(df[df["architecture"] == architecture]).dropna(subset=required_cols).copy()
 
     if df_clean.empty:
-        print(f"No data for architecture '{architecture}'")
+        logger.warning(f"No data for architecture '{architecture}'")
         return
 
     df_clean["log_loss"] = np.log2(df_clean[metric])
@@ -603,6 +515,433 @@ def visualize_loss_by_param_and_epoch_count(
     save_figure(fig, output_path)
 
 
+# ------------------------------------------------------------
+# Isoflop scaling law analysis (from exp2101_plantcad_isoflop_fit.py)
+# ------------------------------------------------------------
+
+
+def fit_quadratic_optimum(x_vals, loss_vals):
+    """
+    Fits L = a*(ln x)^2 + b*(ln x) + c and returns optimal x.
+    Returns (x_opt, fit_coeffs).
+    Raises ValueError if the fit is concave (no minimum).
+    """
+    log_x = np.log(x_vals)
+    coeffs = np.polyfit(log_x, loss_vals, 2)  # [a, b, c]
+    a, b, _ = coeffs
+
+    if a <= 0:
+        raise ValueError(f"Concave fit detected: {coeffs}")
+
+    # Minimum at ln x = -b / (2a)
+    ln_x_opt = -b / (2 * a)
+    x_opt = np.exp(ln_x_opt)
+    return x_opt, coeffs
+
+
+def analyze_budgets(df):
+    """
+    Analyzes each budget group to find optimal N and D using independent quadratic fits.
+    Returns a DataFrame with columns: budget, opt_N, opt_D, coeffs_N, coeffs_D, group_data.
+    """
+    budgets = sorted(df["flops_budget"].unique())
+    results = []
+
+    for budget in budgets:
+        group = df[df["flops_budget"] == budget].sort_values("params")
+        if len(group) < 3:
+            raise ValueError(f"Budget {budget} has fewer than 3 points ({len(group)}), cannot fit.")
+
+        N = group["params"].values
+        D = group["tokens"].values
+        L = group["eval_loss"].values
+
+        # Fit independent quadratics
+        opt_N, coeffs_N = fit_quadratic_optimum(N, L)
+        opt_D, coeffs_D = fit_quadratic_optimum(D, L)
+
+        results.append(
+            {
+                "budget": budget,
+                "opt_N": opt_N,
+                "opt_D": opt_D,
+                "coeffs_N": coeffs_N,
+                "coeffs_D": coeffs_D,
+                "group_data": group,
+            }
+        )
+
+    return pd.DataFrame(results)
+
+
+def fit_scaling_law(budgets, optimal_vals):
+    """
+    Fits log(optimal_val) = m * log(budget) + c.
+    Returns (m, c, B_smooth, V_smooth) where B_smooth and V_smooth are smoothed
+    budget and predicted value arrays for plotting.
+    """
+    log_B = np.log(budgets)
+    log_V = np.log(optimal_vals)
+
+    m, c = np.polyfit(log_B, log_V, 1)
+
+    # Generate smooth line
+    B_smooth = np.logspace(np.log10(budgets.min()), np.log10(budgets.max()), 100)
+    V_smooth = np.exp(m * np.log(B_smooth) + c)
+
+    return m, c, B_smooth, V_smooth
+
+
+def plot_isoflop_curves(ax_N, ax_D, analysis_results, colors_N, colors_D):
+    """Plots the top row: Loss vs Params/Tokens for each budget."""
+    for idx, row in analysis_results.iterrows():
+        budget = row["budget"]
+
+        # Determine color based on index
+        # Assuming colors_N and colors_D are arrays matching analysis_results length
+        color_n = colors_N[idx]
+        color_d = colors_D[idx]
+
+        group = row["group_data"]
+
+        # Plot Data - N (Blues)
+        ax_N.scatter(group["params"], group["eval_loss"], color=color_n, alpha=0.7, s=20, label=f"{budget:.1e}")
+
+        # Plot Data - D (Greens)
+        ax_D.scatter(group["tokens"], group["eval_loss"], color=color_d, alpha=0.7, s=20, label=f"{budget:.1e}")
+
+        # Plot Fits
+        N_range = np.logspace(np.log10(group["params"].min() * 0.5), np.log10(group["params"].max() * 2.0), 100)
+        D_range = np.logspace(np.log10(group["tokens"].min() * 0.5), np.log10(group["tokens"].max() * 2.0), 100)
+
+        L_pred_N = np.polyval(row["coeffs_N"], np.log(N_range))
+        L_pred_D = np.polyval(row["coeffs_D"], np.log(D_range))
+
+        # Lines use the same color scale
+        ax_N.plot(N_range, L_pred_N, color=color_n, linestyle="--", alpha=0.5)
+        ax_D.plot(D_range, L_pred_D, color=color_d, linestyle="--", alpha=0.5)
+
+        # Plot Optima
+        L_min_N = np.polyval(row["coeffs_N"], np.log(row["opt_N"]))
+        ax_N.scatter([row["opt_N"]], [L_min_N], color=color_n, marker="s", s=100, edgecolors="black", zorder=10)
+
+        L_min_D = np.polyval(row["coeffs_D"], np.log(row["opt_D"]))
+        ax_D.scatter([row["opt_D"]], [L_min_D], color=color_d, marker="s", s=100, edgecolors="black", zorder=10)
+
+    # Configure axes
+    for ax, xlabel, title in [(ax_N, "Parameters (N)", "Loss vs Parameters"), (ax_D, "Tokens (D)", "Loss vs Tokens")]:
+        ax.set_xscale("log")
+        ax.set_xlabel(xlabel)
+        ax.set_ylabel("Validation Loss")
+        ax.set_title(title)
+        ax.grid(True, which="both", ls="-", alpha=0.2, axis="x")
+        # Don't show legend here anymore, global legend instead
+
+
+def plot_scaling_laws(ax_N, ax_D, analysis_results, colors_N, colors_D):
+    """Plots the bottom row: Optimal Params/Tokens vs FLOPs."""
+    valid_data = analysis_results.dropna(subset=["opt_N", "opt_D"])
+    valid_indices = valid_data.index
+
+    # Extract colors corresponding to valid data points
+    valid_colors_N = colors_N[valid_indices]
+    valid_colors_D = colors_D[valid_indices]
+
+    budgets = valid_data["budget"].values
+    opt_N = valid_data["opt_N"].values
+    opt_D = valid_data["opt_D"].values
+
+    # Fit Scaling Laws
+    m_N, c_N, B_smooth_N, N_smooth = fit_scaling_law(budgets, opt_N)
+    m_D, c_D, B_smooth_D, D_smooth = fit_scaling_law(budgets, opt_D)
+
+    # Plot Params Scaling
+    ax_N.scatter(budgets, opt_N, color=valid_colors_N, marker="s", s=100, edgecolors="black", zorder=5)
+    ax_N.plot(B_smooth_N, N_smooth, color="gray", linestyle="--", label=f"$N^* \\propto C^{{{m_N:.3f}}}$")
+
+    # Plot Tokens Scaling
+    ax_D.scatter(budgets, opt_D, color=valid_colors_D, marker="s", s=100, edgecolors="black", zorder=5)
+    ax_D.plot(B_smooth_D, D_smooth, color="gray", linestyle="--", label=f"$D^* \\propto C^{{{m_D:.3f}}}$")
+
+    # Configure axes
+    for ax, ylabel, title in [
+        (ax_N, "Optimal Parameters (N*)", "Optimal Parameters vs Compute"),
+        (ax_D, "Optimal Tokens (D*)", "Optimal Tokens vs Compute"),
+    ]:
+        ax.set_xscale("log")
+        ax.set_yscale("log")
+        ax.set_xlabel("Compute Budget (FLOPs)")
+        ax.set_ylabel(ylabel)
+        ax.set_title(title)
+        ax.grid(True, which="both", ls="-", alpha=0.2, axis="x")
+        ax.legend()
+
+    # Calculate Ratio Function
+    ratio_coeff = np.exp(c_D - c_N)
+    ratio_exp_diff = m_D - m_N  # Simple difference of exponents
+    # Optimal ratio exponent: R* = (m_N - m_D) / (m_N + m_D)
+    # This is the Chinchilla optimal token/param ratio exponent
+    ratio_exp_opt = (m_N - m_D) / (m_N + m_D)
+
+    ratio_str = (
+        f"Optimal Ratio: $\\frac{{D_{{opt}}}}{{N_{{opt}}}} = "
+        f"{ratio_exp_opt:.4f}$ ($\\frac{{D^*}}{{N^*}} = {ratio_exp_diff:.4f}$)"
+    )
+
+    # Return all computed values for summary display
+    return {
+        "ratio_str": ratio_str,
+        "m_N": m_N,
+        "c_N": c_N,
+        "m_D": m_D,
+        "c_D": c_D,
+        "ratio_coeff": ratio_coeff,
+        "ratio_exp_diff": ratio_exp_diff,
+        "ratio_exp_opt": ratio_exp_opt,
+    }
+
+
+class DualColorMarker:
+    """
+    Custom marker handler for legend that draws two markers side-by-side.
+    """
+
+    def __init__(self, color1, color2):
+        self.color1 = color1
+        self.color2 = color2
+
+    def legend_artist(self, legend, orig_handle, fontsize, handlebox):
+        x0, y0 = handlebox.xdescent, handlebox.ydescent
+        width, height = handlebox.width, handlebox.height
+
+        # Increase marker size (taller and wider)
+        rect_h = height * 1.4  # Significantly larger
+        rect_w = rect_h * 1.2  # Aspect ratio ~ 1.2 (wide)
+
+        # Centered vertically
+        y_pos = y0 + (height - rect_h) / 2
+
+        # Gap between markers
+        marker_gap = 5  # Fixed pixel gap
+
+        # Total width of the dual marker group
+        total_group_width = rect_w + marker_gap + rect_w
+
+        # Align the dual marker group to the RIGHT side of the handlebox area
+        # This pushes it closer to the text label which starts immediately after handlebox
+        start_x = x0 + width - total_group_width
+
+        # Draw two markers side-by-side
+        # Left marker (Params/Blue)
+        p1 = plt.Rectangle(
+            [start_x, y_pos],
+            rect_w,
+            rect_h,
+            facecolor=self.color1,
+            edgecolor="black",
+            transform=handlebox.get_transform(),
+        )
+        # Right marker (Tokens/Green)
+        p2 = plt.Rectangle(
+            [start_x + rect_w + marker_gap, y_pos],
+            rect_w,
+            rect_h,
+            facecolor=self.color2,
+            edgecolor="black",
+            transform=handlebox.get_transform(),
+        )
+
+        handlebox.add_artist(p1)
+        handlebox.add_artist(p2)
+        return [p1, p2]
+
+
+def plot_scaling_extrapolation(analysis_results, scaling_results, out_dir):
+    """Creates a figure showing N*, D*, and D*/N* vs compute with extrapolation."""
+    valid_data = analysis_results.dropna(subset=["opt_N", "opt_D"])
+    budgets = valid_data["budget"].values
+    opt_N, opt_D = valid_data["opt_N"].values, valid_data["opt_D"].values
+
+    m_N, c_N = scaling_results["m_N"], scaling_results["c_N"]
+    m_D, c_D = scaling_results["m_D"], scaling_results["c_D"]
+    ratio_exp_diff = scaling_results["ratio_exp_diff"]  # Simple: m_D - m_N
+    ratio_exp_opt = scaling_results["ratio_exp_opt"]  # R*: (m_N - m_D) / (m_N + m_D)
+
+    # Extrapolate from min observed to 1e22 FLOPs
+    C_ext = np.logspace(np.log10(budgets.min()), 22, 200)
+    N_ext = np.exp(c_N) * C_ext**m_N
+    D_ext = np.exp(c_D) * C_ext**m_D
+    # Use R* (optimal exponent) for the ratio line
+    # D*/N* = ratio_coeff * C^(-R*) since R* = (m_N - m_D)/(m_N + m_D) and we want D*/N*
+    ratio_coeff = scaling_results["ratio_coeff"]
+    ratio_ext = ratio_coeff * C_ext ** (-ratio_exp_opt)
+
+    _, ax1 = plt.subplots(figsize=(10, 4.8))
+    ax1.set_xscale("log")
+    ax1.set_yscale("log")
+
+    # Plot fits and data
+    (ln,) = ax1.plot(C_ext, N_ext, color="tab:blue", lw=2, label=f"N* ∝ C^{m_N:.3f}")
+    (ld,) = ax1.plot(C_ext, D_ext, color="tab:green", lw=2, label=f"D* ∝ C^{m_D:.3f}")
+    ax1.scatter(budgets, opt_N, color="tab:blue", s=80, edgecolors="black", zorder=5, marker="o")
+    ax1.scatter(budgets, opt_D, color="tab:green", s=80, edgecolors="black", zorder=5, marker="s")
+
+    # Extrapolation shading
+    ax1.axvspan(budgets.max(), 1e22, alpha=0.1, color="gray")
+    ax1.axvline(budgets.max(), color="gray", ls=":", alpha=0.5)
+    ax1.set_xlabel("Compute Budget C (FLOPs)", fontsize=12)
+    ax1.set_ylabel("Optimal N* (params) / D* (tokens)", fontsize=12)
+    ax1.grid(True, which="major", ls="-", alpha=0.2)
+
+    # Right axis: Ratio (set behind ax1 so annotations render on top)
+    ax2 = ax1.twinx()
+    ax2.set_yscale("log")
+    ax2.set_zorder(ax1.get_zorder() - 1)
+    ax1.patch.set_visible(False)
+    # Label shows both simple and normalized exponents
+    (lr,) = ax2.plot(
+        C_ext,
+        ratio_ext,
+        color="gray",
+        lw=2,
+        ls="--",
+        label=f"D*/N* (R*={ratio_exp_opt:.3f}, Δm={ratio_exp_diff:.3f})",
+    )
+    ax2.set_ylabel("Optimal Ratio D*/N*", fontsize=12)
+
+    # Combined legend
+    ax1.legend(handles=[ln, ld, lr], loc="upper left")
+
+    # Reference annotations
+    for C_ref in [1e18, 1e20, 1e22]:
+        N_ref, D_ref = np.exp(c_N) * C_ref**m_N, np.exp(c_D) * C_ref**m_D
+        ax1.axvline(C_ref, color="gray", ls=":", alpha=0.3)
+        ax1.annotate(
+            f"C={C_ref:.0e}\nN*={N_ref:.1e}\nD*={D_ref:.1e}\nD*/N*={D_ref/N_ref:.0f}",
+            xy=(C_ref, N_ref * 1.5),
+            fontsize=8,
+            ha="center",
+            va="bottom",
+            zorder=20,
+            bbox=dict(boxstyle="round,pad=0.3", facecolor="white", alpha=0.8, edgecolor="gray"),
+        )
+
+    plt.title("Scaling Law Extrapolation: Optimal Compute Allocation", fontsize=14)
+    plt.tight_layout()
+
+    out_png, out_pdf = out_dir / "plantcad_scaling_extrapolation.png", out_dir / "plantcad_scaling_extrapolation.pdf"
+    plt.savefig(out_png, dpi=300, bbox_inches="tight")
+    plt.savefig(out_pdf, dpi=300, bbox_inches="tight")
+    plt.close()
+    return out_png, out_pdf
+
+
+def print_summary(df, analysis_results, scaling_results, out_files):
+    """Print summary using rich tables."""
+    console.print()
+
+    # Optimal allocations table
+    opt_table = Table(title="Optimal Allocations")
+    opt_table.add_column("Budget (C)", justify="right", style="cyan")
+    opt_table.add_column("Opt N*", justify="right")
+    opt_table.add_column("Opt D*", justify="right")
+    opt_table.add_column("D*/N*", justify="right", style="yellow")
+
+    for _, row in analysis_results.iterrows():
+        ratio = row["opt_D"] / row["opt_N"] if row["opt_N"] > 0 else np.nan
+        opt_table.add_row(
+            f"{row['budget']:.1e}",
+            f"{row['opt_N']:.2e}",
+            f"{row['opt_D']:.2e}",
+            f"{ratio:.1f}",
+        )
+    console.print(opt_table)
+    console.print()
+
+    # Scaling laws
+    m_N, c_N = scaling_results["m_N"], scaling_results["c_N"]
+    m_D, c_D = scaling_results["m_D"], scaling_results["c_D"]
+    ratio_coeff = scaling_results["ratio_coeff"]
+    ratio_exp_diff = scaling_results["ratio_exp_diff"]
+    ratio_exp_opt = scaling_results["ratio_exp_opt"]
+
+    console.print("[bold]Scaling Laws:[/bold]")
+    console.print(f"  N* ∝ C^{m_N:.3f}  [dim](log N* = {m_N:.4f} log C + {c_N:.4f})[/dim]")
+    console.print(f"  D* ∝ C^{m_D:.3f}  [dim](log D* = {m_D:.4f} log C + {c_D:.4f})[/dim]")
+    console.print(f"  D*/N* = {ratio_coeff:.4e} · C^{ratio_exp_diff:.4f}")
+    console.print(f"  R* = (m_N - m_D)/(m_N + m_D) = {ratio_exp_opt:.4f}")
+    console.print()
+
+    for f in out_files:
+        console.print(f"[green]✓[/green] Saved: {f}")
+    console.print()
+
+
+def run_isoflop_fit_analysis(df: pd.DataFrame, architecture: str = DEFAULT_ARCH) -> None:
+    """Run isoflop scaling law fitting and visualization on epoch=1 data for the given architecture."""
+    df_fit = filter_to_finished_runs(df)
+    df_fit = df_fit[(df_fit["architecture"] == architecture) & (df_fit["epochs"] == 1)].dropna(
+        subset=["eval_loss", "tokens", "params", "flops_budget"]
+    )
+    if df_fit.empty:
+        raise ValueError("No valid data found for isoflop fitting")
+    analysis_results = analyze_budgets(df_fit)
+
+    # Setup Figure: 2x2 Grid
+    fig, axes = plt.subplots(2, 2, figsize=(12, 8), gridspec_kw={"height_ratios": [1.5, 1]})
+
+    # Colorscales
+    colors_N = plt.cm.Blues(np.linspace(0.4, 1.0, len(analysis_results)))
+    colors_D = plt.cm.Greens(np.linspace(0.4, 1.0, len(analysis_results)))
+
+    # Top Row
+    plot_isoflop_curves(axes[0, 0], axes[0, 1], analysis_results, colors_N, colors_D)
+
+    # Bottom Row
+    scaling_results = plot_scaling_laws(axes[1, 0], axes[1, 1], analysis_results, colors_N, colors_D)
+
+    # Global Legend with dual-color markers
+    legend_handles = []
+    for _, row in analysis_results.iterrows():
+        legend_handles.append(plt.Rectangle((0, 0), 1, 1, color="none", label=f"{row['budget']:.1e}"))
+
+    handler_map = {legend_handles[i]: DualColorMarker(colors_N[i], colors_D[i]) for i in range(len(legend_handles))}
+
+    leg = fig.legend(
+        handles=legend_handles,
+        handler_map=handler_map,
+        title="Compute Budget [$C$]\n(FLOPs)",
+        loc="center left",
+        bbox_to_anchor=(0.86, 0.5),
+        borderaxespad=0.5,
+        handlelength=2.5,
+        labelspacing=0.6,
+        handletextpad=0.8,
+    )
+    leg.get_title().set_multialignment("center")
+
+    plt.suptitle("PlantCAD Isoflop Analysis: Quadratic Optima & Scaling Laws", y=0.94, fontsize=16)
+    plt.figtext(0.5, 0.88, scaling_results["ratio_str"], ha="center", fontsize=14)
+    plt.tight_layout(rect=[0, 0, 0.85, 0.91])
+
+    out_dir = Path(RESULT_PATH)
+    out_dir.mkdir(parents=True, exist_ok=True)
+    out_png = out_dir / "plantcad_isoflop_fits.png"
+    out_pdf = out_dir / "plantcad_isoflop_fits.pdf"
+    plt.savefig(out_png, dpi=300, bbox_inches="tight")
+    plt.savefig(out_pdf, dpi=300, bbox_inches="tight")
+    plt.close()
+
+    # Create extrapolation figure
+    extrap_png, extrap_pdf = plot_scaling_extrapolation(analysis_results, scaling_results, out_dir)
+
+    # Print summary
+    out_files = [out_png, out_pdf, extrap_png, extrap_pdf]
+    print_summary(df_fit, analysis_results, scaling_results, out_files)
+
+
+# ------------------------------------------------------------
+
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Fetch and analyze plantcad isoflop runs")
     parser.add_argument(
@@ -622,15 +961,19 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
 
+    # Setup logging to console and file
+    log_path = Path(RESULT_PATH) / "plantcad_isoflop_analysis.txt"
+    setup_logging(log_path)
+
     output_path = Path(args.output)
 
     # Check if CSV exists and load from it unless --force is specified
     if output_path.exists() and not args.force:
-        print(f"Loading existing data from {output_path}")
+        logger.info(f"Loading existing data from {output_path}")
         df = pd.read_csv(output_path)
-        print(f"Loaded {len(df)} runs from CSV")
+        logger.info(f"Loaded {len(df)} runs from CSV")
     else:
-        print("Fetching runs from W&B...")
+        logger.info("Fetching runs from W&B...")
         df = fetch_plantcad_runs(show_wandb_runs=args.show_wandb_runs)
         save_runs(df, output_path)
 
@@ -639,5 +982,12 @@ if __name__ == "__main__":
     visualize_loss_by_token_count(df)
     visualize_loss_by_param_count(df)
     visualize_loss_by_epochs(df)
-    visualize_loss_by_epochs_matplotlib(df)
     visualize_loss_by_param_and_epoch_count(df)
+    run_isoflop_fit_analysis(df)
+
+    # Append rich console output to log file
+    with open(log_path, "a") as f:
+        f.write("\n" + console.export_text())
+        f.write(f"\nAnalysis complete. Logs saved to {log_path}\n")
+    # Also print to console
+    console.print(f"[green]Analysis complete.[/green] Logs saved to {log_path}")
